+++ plugins/aap-backend/src/providers/AapResourceEntityProvider.ts
@@ -1,3 +1,4 @@
+import { LoggerService } from '@backstage/backend-plugin-api';
 import { PluginTaskScheduler, TaskRunner } from '@backstage/backend-tasks';
 import {
   ANNOTATION_LOCATION,
@@ -10,8 +11,6 @@ import {
   EntityProvider,
   EntityProviderConnection,
 } from '@backstage/plugin-catalog-node';
-
-import { Logger } from 'winston';
 
 import {
   listJobTemplates,
@@ -27,14 +26,14 @@ export class AapResourceEntityProvider implements EntityProvider {
   private readonly authorization: string;
   private readonly owner: string;
   private readonly system: string;
-  private readonly logger: Logger;
+  private readonly logger: LoggerService;
   private readonly scheduleFn: () => Promise<void>;
   private connection?: EntityProviderConnection;
 
   static fromConfig(
     configRoot: Config,
     options: {
-      logger: Logger;
+      logger: LoggerService;
       schedule?: TaskRunner;
       scheduler?: PluginTaskScheduler;
     },
@@ -65,7 +64,7 @@ export class AapResourceEntityProvider implements EntityProvider {
 
   private constructor(
     config: AapConfig,
-    logger: Logger,
+    logger: LoggerService,
     taskRunner: TaskRunner,
   ) {
     this.env = config.id;