+++ plugins/servicenow-actions/CHANGELOG.md
@@ -1,3 +1,26 @@
+## @janus-idp/backstage-scaffolder-backend-module-servicenow [1.4.15](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.14...@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.15) (2024-06-13)
+
+
+### Bug Fixes
+
+* missing postversion script is several plugins and missing turbo dependency ([#1811](https://github.com/janus-idp/backstage-plugins/issues/1811)) ([4dfe4f5](https://github.com/janus-idp/backstage-plugins/commit/4dfe4f533e21e79c928c66bfd68684243912be2c))
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.10.1
+
+## @janus-idp/backstage-scaffolder-backend-module-servicenow [1.4.14](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.13...@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.14) (2024-06-05)
+
+
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.10.0
+
+## @janus-idp/backstage-scaffolder-backend-module-servicenow [1.4.13](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.12...@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.13) (2024-06-04)
+
 ## @janus-idp/backstage-scaffolder-backend-module-servicenow [1.4.12](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.11...@janus-idp/backstage-scaffolder-backend-module-servicenow@1.4.12) (2024-06-03)